requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
